Sincronized with main and implemented informations from receive in connection
