+++ core/settings.py
@@ -35,7 +35,6 @@ ALLOWED_HOSTS = env('ALLOWED_HOSTS').split(' ')
 
 # Application definition
 INSTALLED_APPS = [
-    'daphne',
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
@@ -46,10 +45,8 @@ INSTALLED_APPS = [
     'rest_framework',
     'corsheaders',
     'knox',
-    'channels',
     # Apps
     'apps.user',
-    'apps.chat',
     'apps.shapeteam',
 ]
 
@@ -84,7 +81,6 @@ TEMPLATES = [
     },
 ]
 
-ASGI_APPLICATION = 'core.asgi.application'
 WSGI_APPLICATION = 'core.wsgi.application'
 
 # Database
@@ -99,16 +95,6 @@ if DEBUG:
 else:
     import dj_database_url
     DATABASES = {'default': dj_database_url.parse(env('DATABASE_URL'))}
-
-
-CHANNEL_LAYERS = {
-    "default": {
-        "BACKEND": "channels_redis.core.RedisChannelLayer",
-        "CONFIG": {
-            "hosts": [("127.0.0.1", 6379)],
-        },
-    },
-}
 
 # Password validation
 # https://docs.djangoproject.com/en/5.0/ref/settings/#auth-password-validators
