+++ requirements.txt
@@ -5,11 +5,9 @@ autobahn==24.4.2
 Automat==24.8.1
 Brotli==1.1.0
 cffi==1.17.1
-channels==4.2.0
 click==8.1.7
 constantly==23.10.4
 cryptography==44.0.0
-daphne==4.1.2
 decorator==5.1.1
 dj-database-url==2.2.0
 Django==5.0.2
